Implement AO Network client and message handling
- Introduced the AO client module, including the `AOClient` trait and its production implementation.
- Added configuration management with `AOConfig` for network connection settings.
- Defined message types for AO communication in the `message` module, facilitating interaction with AO processes.
- Implemented a mock client for testing purposes, allowing for development without actual network connections.
- Enhanced the existing adapter structure to support the new AO functionalities, ensuring a cohesive integration with the external system adapters.
