+++ client/src/adapter/external/mock_ao/mod.rs
@@ -1,0 +1,8 @@
+//! Mock AO Network client module
+//!
+//! Provides mock implementation of AOClient trait
+//! for testing and development.
+
+mod client;
+
+pub use client::{MockAOClient, MockConfig};
